[ambipi] Splitting up source files
